Refactored datasets package
